Move Unknown one namespace level up
