+++ contribution/generator/src/TrackData/CanonicalData.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace App\TrackData;
 
 use App\TrackData\TestCase;
-use App\TrackData\CanonicalData\Unknown;
+use App\TrackData\Unknown;
 
 class CanonicalData
 {
